Poll_dapp-main/requirements.txt
asn1crypto==0.24.0
attrs==20.3.0
Automat==0.6.0
backcall==0.2.0
base58==2.0.1
bitarray==1.2.2
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==6.7
cloud-init==20.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cytoolz==0.11.0
decorator==4.4.2
distro-info===0.18ubuntu0.18.04.1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
hexbytes==0.2.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
importlib-metadata==3.4.0
importlib-resources==4.1.1
incremental==16.10.1
ipfshttpclient==0.7.0a1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter-core==4.7.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lru-dict==1.1.6
MarkupSafe==1.0
multiaddr==0.0.9
nbformat==5.0.8
netaddr==0.8.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
parsimonious==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pigar==0.10.0
prompt-toolkit==3.0.9
protobuf==3.14.0
ptyprocess==0.7.0
py-solc==3.2.0
py-solc-x==1.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycryptodome==3.9.9
Pygments==2.7.3
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyrsistent==0.17.3
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.25.1
requests-unixsocket==0.1.5
rlp==2.0.1
SecretStorage==2.3.1
semantic-version==2.8.5
service-identity==16.0.0
six==1.15.0
ssh-import-id==5.7
systemd-python==234
toolz==0.11.1
traitlets==4.3.3
Twisted==17.9.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
varint==1.0.2
wcwidth==0.2.5
web3==5.13.1
websockets==8.1
zipp==3.4.0
zope.interface==4.3.2
